move cheatcode abstract class to starkware dir
