+++ protostar/commands/test/cheatcodes/deploy_cheatcode.py
@@ -5,8 +5,8 @@ from typing import Any, Callable
 from starkware.python.utils import to_bytes
 from starkware.starknet.core.os.syscall_utils import initialize_contract_state
 
-from protostar.commands.test.cheatcodes.cheatcode import Cheatcode
 from protostar.commands.test.cheatcodes.prepare_cheatcode import PreparedContract
+from protostar.commands.test.starkware.cheatcode import Cheatcode
 
 
 @dataclass(frozen=True)
