+++ protostar/commands/test/cheatcodes/prepare_cheatcode.py
@@ -6,8 +6,8 @@ from starkware.starknet.core.os.contract_address.contract_address import (
     calculate_contract_address_from_hash,
 )
 
-from protostar.commands.test.cheatcodes.cheatcode import Cheatcode
 from protostar.commands.test.cheatcodes.declare_cheatcode import DeclaredContract
+from protostar.commands.test.starkware.cheatcode import Cheatcode
 from protostar.commands.test.test_environment_exceptions import CheatcodeException
 from protostar.utils.data_transformer_facade import DataTransformerFacade
