+++ protostar/commands/test/starkware/__init__.py
@@ -1,9 +1,9 @@
+from protostar.commands.test.starkware.cheatable_starknet_general_config import (
+    CheatableStarknetGeneralConfig,
+)
 from protostar.commands.test.starkware.cheatable_syscall_handler import (
     CheatableHintsWhitelist,
     CheatableSysCallHandler,
     CheatableSysCallHandlerException,
-)
-from protostar.commands.test.starkware.cheatable_starknet_general_config import (
-    CheatableStarknetGeneralConfig,
 )
 from protostar.commands.test.starkware.forkable_starknet import ForkableStarknet
