+++ protostar/commands/test/starkware/cheatcode.py
